Start query refactor.
